+++ CHANGELOG.md
@@ -11,6 +11,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 * Correctly filter Xcode warnings about the use of dynamic frameworks.  
   [Boris Bügling](https://github.com/neonichu)
 
+* Lint to prevent dynamic libraries and frameworks from passing with iOS 7.  
+  [Boris Bügling](https://github.com/neonichu)
+  [#3193](https://github.com/CocoaPods/CocoaPods/issues/3193)
+
 
 ## 0.36.0.rc.1